requirements.txt
absl-py==0.10.0
appdirs==1.4.4
astunparse==1.6.3
black==20.8b1
bleach==3.2.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
docformatter==1.3.1
docutils==0.16
flake8==3.8.3
future==0.18.2
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.33.1
h5py==2.10.0
idna==2.10
joblib==0.17.0
Keras-Preprocessing==1.1.2
keyring==21.4.0
Markdown==3.3.2
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.1
pathspec==0.8.0
pkginfo==1.6.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
readme-renderer==28.0
regex==2020.10.15
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.3
six==1.15.0
sparsemax==0.1.9
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.1
torch==1.6.0
tqdm==4.50.2
twine==3.2.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
untokenize==0.1.1
urllib3==1.25.11
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xgboost==1.2.1
neptune-client==0.9.18
